+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `filament-js-money-field` will be documented in this file.
 
+## 0.1.0 - 2024-12-09
+
+### Added
+- Support for `decimal`, `double`, and `float` column types, in addition to the already supported `integer`.
+
+### Fixed
+- Formatting issues in `column` and `entry`.
+
 ## 0.0.1 - 2024-11-28
 
 - initial release